Move get_flex_image* to their own module
This probably still isn't a great place, but means that we can
import them without importing wx (which makes working wx a dependency
in CI systems)
